Automatic request mapping do not happen for port 8080
On the note section in the following link - https://docs.microsoft.com/en-us/azure/java/spring-framework/deploy-spring-boot-java-app-on-linux#create-a-web-app-on-linux-on-azure-app-service-using-your-container-image , it is stated that "Azure will automatically map Internet requests to embedded Tomcat server that is running on the standard ports of 80 or 8080. "

But this seems to be not working. The automatic mapping happens for just port 80 and for 8080, the mapping needs to be configured in Application settings using WEBSITES_PORT setting.

Closes #80