Added the 'depends_on' Meta-Argument for kubesubnet and appgwsubnet
Added the 'depends_on' Meta-Argument for azurerm_subnet.kubesubnet and azurerm_subnet.appgwsubnet. This is done to prevent the below error:

Error: Subnet "kubesubnet" was not found
Error: Subnet "appgwsubnet" was not found 

The depends_on Meta-Argument will ensure that the VNet and both the subnets are created first and then the information is read.