Clarify steps to test node function locally
When it says this is a "reminder", it makes me go back to find where I missed a step - but it doesn't seem that is specified anywhere else when going through this material, so it ends up being a waste of time.  This puts the steps in proper order for someone going step by step through this tutorial.