Update connect-from-azure.md
The service principal does not need full contributor rights on the subscription itself. Documentation needs to clarify that it's possible to assign contributor rights to the resources (i.e. resource group) for the service principal to perform actions. Ideally we will apply the principle of least privilege when granting permissions.