Incorrect link github-key-vault.md
Hi.

This "broken" link was "fixed" incorrectly a while ago; probably when the GitHub Action it was referencing was deprecated and likely removed from the Marketplace. However it was replaced by some third-party non-Microsoft Enhanced ENV Azure Vault Action that works quite differently.

Instead of pointing to a missing Marketplace page, I'm updating the link to the action's GitHub repo that contains the docs and code.