Policy Signing Certificate has to be an X.509 Certificate
A PEM file that is obtained when creating a VM does only contain a private key. However, MAA requires a public key to validate signed policies, which is provided by a signed or self-signed X.509 certificate. If you try to create an attestation provider by using an SSH private key (a PEM file that contains a private key), as suggested in the article, the provisioning fails with this error: "PolicySigningCertificate key is not an X509 security key".

Therefore I suggest these updates to meet the requirements to launch an attestation provider.